Update to use storybooks docs gen
