+++ src/components/dropdown-button/index.tsx
@@ -1,84 +1,3 @@
-import React, { useMemo, useRef } from 'react';
-import { Flex, FlexProps } from 'rebass';
-import * as R from 'ramda';
+export { DropdownButton } from './DropdownButton';
 
-import styles from './dropdown-button.styles';
-import useDropdown from '../../utils/useDropdown';
-import Dropdown, { DropdownProps } from '../dropdown';
-import { DropdownItem } from '../dropdown/types';
-
-type RenderDropdownItems = ({
-  onClose,
-}: {
-  onClose: () => void;
-}) => DropdownProps['items'];
-export interface DropdownButtonProps extends Omit<FlexProps, 'css'> {
-  /** wether to align dropdown on the left or right, the default is right alignment
-   * TODO: in the future we might need excact place of opening dropdown similar to tooltip
-   */
-  alignLeft?: boolean;
-  containerOpenSx?: FlexProps['sx'];
-  dropdownSx?: DropdownProps['sx'];
-  /** dropwdown items weather could be an array, in this case dropdown will be closed after any item clicked
-   *  in order to handle closing of dropdown manually you have to pass a callback function which accepts the onClose handler function
-   */
-  items: DropdownProps['items'] | RenderDropdownItems;
-  renderButton: ({
-    onClick,
-    isOpen,
-  }: {
-    onClick: () => void;
-    isOpen: boolean;
-  }) => React.ReactNode;
-}
-
-const DropdownButton: React.FC<DropdownButtonProps> = ({
-  alignLeft = false,
-  containerOpenSx = {},
-  dropdownSx = {},
-  items,
-  renderButton,
-  sx = {},
-  ...restProps
-}) => {
-  const [isOpen, handleToggle, handleClickOutside] = useDropdown();
-  const containerRef = useRef<HTMLElement>();
-  const dropdownItems = useMemo(
-    () =>
-      R.equals(R.type(items), 'Array')
-        ? /** if the items is array then close dropdown after clicking each dropdown item */
-          (items as DropdownItem[]).map((item) => ({
-            ...item,
-            onClick: () => {
-              item.onClick(item);
-              handleToggle();
-            },
-          }))
-        : (items as RenderDropdownItems)({ onClose: handleToggle }),
-    [items, handleToggle],
-  );
-
-  return (
-    <Flex
-      ref={containerRef}
-      sx={R.mergeDeepRight(styles.container, {
-        ...sx,
-        ...(isOpen ? containerOpenSx : {}),
-      })}
-      {...restProps}
-    >
-      {renderButton({ isOpen, onClick: handleToggle })}
-      <Dropdown
-        isOpen={isOpen}
-        onClickOutside={handleClickOutside}
-        sx={R.mergeDeepRight(
-          styles.dropdown(containerRef?.current?.offsetHeight, alignLeft),
-          { ...dropdownSx },
-        )}
-        items={dropdownItems}
-      />
-    </Flex>
-  );
-};
-
-export default DropdownButton;
+export type { Props as UserProps } from './DropdownButton';
